+++ apps/api/src/app/order/order.module.ts
@@ -4,6 +4,7 @@ import { ConfigurationService } from '../../services/configuration.service';
 import { DataGatheringService } from '../../services/data-gathering.service';
 import { DataProviderService } from '../../services/data-provider.service';
 import { AlphaVantageService } from '../../services/data-provider/alpha-vantage/alpha-vantage.service';
+import { GhostfolioScraperApiService } from '../../services/data-provider/ghostfolio-scraper-api/ghostfolio-scraper-api.service';
 import { RakutenRapidApiService } from '../../services/data-provider/rakuten-rapid-api/rakuten-rapid-api.service';
 import { YahooFinanceService } from '../../services/data-provider/yahoo-finance/yahoo-finance.service';
 import { ImpersonationService } from '../../services/impersonation.service';
@@ -22,6 +23,7 @@ import { OrderService } from './order.service';
     ConfigurationService,
     DataGatheringService,
     DataProviderService,
+    GhostfolioScraperApiService,
     ImpersonationService,
     OrderService,
     PrismaService,